// src/app/interfaces/userPaciente.interface.ts
export interface UserPaciencte {
    apellidos: string;
    cedula: string;
    ciudad: string;
    comentario: string;
    direccion: string;
    email: string;
    fechaEntrega: string;
    medicamento1: string;
    nombres: string;
    numeroMedicamentos: string;
    periodicidadEntrega: string;
    telefono: string;
    tipoEntrega: string;
    tipoId: string;
    tipoRegistro: string;
    tratamientoDatos: string;
  }
  