// src/app/interfaces/userAcudiente.interface.ts
export interface UserAcudiente {
    id: string;
    apellidosAcudiente: string;
    ciudadAcudiente: string;
    emailAcudiente: string;
    nombresAcudiente: string;
    direccionAcudiente: string;
    nombresPaciente: string;
    apellidosPaciente: string;
    tipoIdPaciente: string;
    numeroIdPaciente: string;
    telefonoAcudiente: string;
    tratamientoDatos: string;
    valTratamientoDatos: string;
    comentario: string;
  }